backend/requirements.txt
rasa==3.6.15
rasa-sdk>=3.6.2,<3.7.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
python-socketio==5.11.1
python-engineio==4.9.0
SQLAlchemy>=1.4.0,<1.5.0
pymongo>=3.8,<4.4
redis>=4.5.3,<5.0
flask==3.0.2
flask-cors==4.0.0 